requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
benepar==0.2.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
cymem==2.0.13
datasets==4.5.0
dill==0.4.0
distro==1.9.0
exceptiongroup==1.3.1
filelock==3.24.0
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
openai==2.21.0
pandas==2.3.3
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
sympy==1.14.0
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torch-struct==0.5
tqdm==4.67.3
transformers==5.1.0
typer==0.23.1
typer-slim==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
